Remove logic from AdminControllers

+++ WebApiTest/Services/AdminService/AdminService.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using WebApiTest.Data;
 using WebApiTest.Models;
@@ -67,27 +68,60 @@ namespace WebApiTest.Services.AdminService
         {
             return await _context.Products.FindAsync(id);
         }
-        public async void RemoveProduct(Product productItem)
+        public async Task<ActionResult<ShopDto>> GetShopAsync(int id)
         {
+            var shopItem = await FindShopAsync(id);
+            if (shopItem == null)
+            {
+                return null;
+            }
+            return ShopDto.ItemShopDTO(shopItem);
+        }
+        public async Task<ActionResult<ProductDto>> GetProductAsync(int id)
+        {
+            var productItem = await FindProductAsync(id);
+
+            if (productItem == null)
+            {
+                return null;
+            }
+
+            return ProductDto.ItemProductDTO(productItem);
+        }
+        public async Task<bool> RemoveProduct(int id)
+        {
+            var productItem = await FindProductAsync(id);
+            if (productItem == null)
+            {
+                return false;
+            }
             _context.Products.Remove(productItem);
             await SaveChangesAsync();
-            
+            return true;
         }
 
-        public async void RemoveShop(Shop shopItem)
+        public async Task<bool> RemoveShop(int id)
         {
+            var shopItem = await FindShopAsync(id);
+            if (shopItem == null)
+            {
+                return false;
+            }
             _context.Shops.Remove(shopItem);
             await SaveChangesAsync();
-            
+            return true;
         }
-        public async void AddProduct(Product productItem)
+        public async Task AddProduct(ProductDto productDtoItem)
         {
+            var shop = FindShopAsync(productDtoItem.ShopId);
+            var productItem = productDtoItem.Product(shop.Result);
+
             _context.Products.Add(productItem);
             await SaveChangesAsync();
         }
-        public async void AddShop(Shop shopItem)
+        public async Task AddShop(ShopDto shopDtoItem)
         {
-            var shopDto = _context.Shops.Add(shopItem);
+            var shop = _context.Shops.Add(shopDtoItem.Shop());
             await SaveChangesAsync();
             
         }
